added constants folder to contain all constants, decoupled the screens from components, remove unwanted packages in MapExplorerScreen for sonarQube issues
